backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
torch==2.5.1
torchvision==0.20.1
pillow==11.0.0
sqlalchemy==2.0.36
python-multipart==0.0.20
python-dotenv==1.0.1
google-generativeai==0.8.3
